Apply PCR amplification efficiency estimation to a data.frame

This function applies the est_amplification_efficiency() function to a data.frame object which follows the standardized format shown in the template_WES_standard_curve data set.

Usage

apply_amplification_efficiency(standard_curves)

Arguments

standard_curves

A data.frame giving the target name, serial diluted concentration of target nucleic acid, and Ct value from a standard curve assay. Must follow the template_WES_standard_curve standardized format.

Value

A data.frame containing the mean, and low and high of the 95% confidence interval of the percentile amplification efficiency for each target name.

Examples

apply_amplification_efficiency(template_WES_standard_curve)
#>   target_name      mean    ci_lo     ci_hi
#> 1    target_0 1.0728605 1.376861 0.7688600
#> 2    target_1 1.0970743 1.347295 0.8468532
#> 3    target_2 1.1258847 1.406146 0.8456235
#> 4    target_3 0.9726151 1.222018 0.7232123