Calculate delta delta Ct

This function calculates relative gene expression using the delta delta Ct method described in Livak and Schmittgen (2001). Adjusted delta delta Ct values following Yuan et al. (2008) can be calculated by providing estimated values for the percentile amplification efficiency in pae_* arguments.

Usage

calc_delta_delta_ct(
  ct_target_treatment,
  ct_target_control,
  ct_reference_treatment,
  ct_reference_control,
  pae_target_treatment = 1,
  pae_target_control = 1,
  pae_reference_treatment = 1,
  pae_reference_control = 1
)

Arguments

ct_target_treatment

A numeric scalar providing the Ct value of the target gene for an observation in the treatment group

ct_target_control

A numeric scalar providing the Ct value of the target gene for the reference observation in the control group

ct_reference_treatment

A numeric scalar providing the Ct value of the reference gene for an observation in the treatment group

ct_reference_control

A numeric scalar providing the Ct value of the reference gene for the reference observation in the control group

pae_target_treatment

A numeric scalar providing the percentile amplification efficiency for the target gene and the treatment group. Defaults to 1.

pae_target_control

A numeric scalar providing the percentile amplification efficiency for the target gene and the control group. Defaults to 1.

pae_reference_treatment

A numeric scalar providing the percentile amplification efficiency for the reference gene and the treatment group. Defaults to 1.

pae_reference_control

A numeric scalar providing the percentile amplification efficiency for the reference gene and the control group. Defaults to 1.

Value

Scalar

Examples

# Traditional method
calc_delta_delta_ct(ct_target_treatment = 32.5,
                    ct_reference_treatment = 25,
                    ct_target_control = 34,
                    ct_reference_control = 30)
#> [1] 0.08838835

# Adjusted calculation incorporating amplification efficiency
calc_delta_delta_ct(ct_target_treatment = 32.5,
                    ct_reference_treatment = 25,
                    ct_target_control = 34,
                    ct_reference_control = 30,
                    pae_target_treatment=0.97,
                    pae_target_control=0.98,
                    pae_reference_treatment=0.98,
                    pae_reference_control=0.99)
#> [1] 0.09440454