Get elevation data

This function takes information of where and when a set of environmental samples were collected and retrieves the elevation (in meters) for those locations at an approximate 100m spatial resolution. Data come from the SRTM DEM which are accessed through the Amazon Web Services (AWS) API and the elevatr R package.

Usage

get_elevation_data(lon, lat)

Arguments

lon

A numeric vector giving the longitude of the sampling sites in Decimal Degrees.

lat

A numeric vector giving the latitude of the sampling sites in Decimal Degrees.

Value

data.frame

Examples

# \donttest{

get_elevation_data(lon = template_WES_data$lon,
                   lat = template_WES_data$lat)
#> Total locations = 3
#> Downloading elevation data from AWS API via elevatr package ...
#> Mosaicing & Projecting
#> Note: Elevation units are in meters
#>     lon   lat elevation
#> 1 90.37 23.80        11
#> 2 90.38 23.80        11
#> 3 90.37 23.81        19

# }